requirements.txt
asgiref==3.3.4
coverage==5.5
Django==3.2.3
djangorestframework==3.12.4
numpy==1.20.2
Pillow==8.1.2
psycopg2==2.8.6
psycopg2-binary==2.8.6
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
Unidecode==1.2.0
